broker/requirements.txt
fastapi~=0.89.1
psycopg2-binary
uvicorn